move core interfaces and use the new core interfaces
