--- src/Reservat/Interfaces/EntityInterface.php
@@ -1,8 +1,0 @@
-<?php 
-
-namespace Reservat\Interfaces;
-
-interface EntityInterface
-{
-    public function toArray();
-}
